删除不必要的方法；调整issue创建视图

+++ iOS-git-osc/ProjectDescriptionCell.m
@@ -38,8 +38,8 @@
         _isWatched = isWatched;
         NSString *starAction = isStarred? @"Unstar" : @"Star";
         NSString *watchAction = isWatched? @"Unwatch" : @"Watch";
-        [_starButton setTitle:[NSString stringWithFormat:@" %@ %d ",starAction, starsCount] forState:UIControlStateNormal];
-        [_watchButton setTitle:[NSString stringWithFormat:@" %@ %d ",watchAction, watchesCount] forState:UIControlStateNormal];
+        [_starButton setTitle:[NSString stringWithFormat:@" %@ %ld ",starAction, (long)starsCount] forState:UIControlStateNormal];
+        [_watchButton setTitle:[NSString stringWithFormat:@" %@ %ld ",watchAction, (long)watchesCount] forState:UIControlStateNormal];
         _projectDescriptionField.text = projectDescription.length > 0? projectDescription : @"暂无项目介绍";
     }
